# Optimize recvAndDecompress when receiving compressed payload