The [PR #7356](https://github.com/grpc/grpc-go/pull/7356) in grpc-go introduced a new Codec that uses mem.BufferSlice instead of []byte and also began pooling the underlying []byte slices by default.

With this update, the DecompressedSize method is no longer called for codecs that implement it. Previously, this method was used to determine the size of compressed content, allowing us to allocate the destination []byte with the correct size to avoid oversized slices or reallocation. The reasoning behind the change is that, since the slices are now pooled, the "decompress" method would use multiple smaller slices rather than a single large one, thereby reducing overhead due overallocation/reallocation. See this [comment](https://github.com/grpc/grpc-go/blame/cb329375b14e0ddd4e454e18405a732f7c2ccd72/rpc_util.go#L880-L882).

The problem is that the method [io.Copy](https://github.com/grpc/grpc-go/blame/cb329375b14e0ddd4e454e18405a732f7c2ccd72/rpc_util.go#L903C11-L903C13) allocates a temporary32kb byte slice internally on each call: See

https://github.com/golang/go/blob/889abb17e125bb0f5d8de61bb80ef15fbe2a130d/src/io/io.go#L387-L389

And 

https://github.com/golang/go/blob/889abb17e125bb0f5d8de61bb80ef15fbe2a130d/src/io/io.go#L417-L427

This PR now uses uses the io.CopyBuffer method instead of io.Copy, allowing us to pass a []byte used internally directly as a parameter. This buffer now is also being pulled using the new `mem` package. This change also reintroduces the call to the DecompressedSize method when available or assume the same 32kb to request the the buffer from the pool.

From the [doc](https://pkg.go.dev/io#CopyBuffer)

> CopyBuffer is identical to Copy except that it stages through the provided buffer (if one is required) rather than allocating a temporary one. If buf is nil, one is allocated; otherwise if it has zero length, CopyBuffer panics. 

I create a benchmark with zip (which implements the`DecompressedSize` method) and a `noop` compression that does not implement the `DecompressedSize` method. 

```
goos: linux
goarch: amd64
pkg: google.golang.org/grpc
cpu: Intel(R) Xeon(R) Platinum 8175M CPU @ 2.50GHz
                                              │  /tmp/old   │              /tmp/new              │
                                              │   sec/op    │   sec/op     vs base               │
RPCCompressor/comp=gzip,payloadSize=1024-16     342.0µ ± 3%   295.0µ ± 2%  -13.75% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=10240-16    440.1µ ± 2%   393.9µ ± 2%  -10.49% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=512000-16   5.111m ± 1%   5.179m ± 2%   +1.33% (p=0.004 n=6)
RPCCompressor/comp=noop,payloadSize=1024-16     166.9µ ± 1%   135.3µ ± 1%  -18.93% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=10240-16    183.5µ ± 1%   152.5µ ± 3%  -16.92% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=512000-16   977.7µ ± 2%   941.9µ ± 1%   -3.66% (p=0.002 n=6)
geomean                                         533.4µ        476.4µ       -10.69%

                                              │    /tmp/old    │               /tmp/new               │
                                              │      B/op      │     B/op	vs base               │
RPCCompressor/comp=gzip,payloadSize=1024-16     164.95Ki ± 11%   32.22Ki ±  9%  -80.47% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=10240-16    203.66Ki ±  6%   55.36Ki ± 17%  -72.82% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=512000-16    1.193Mi ± 14%   1.075Mi ±  7%   -9.89% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=1024-16      78.94Ki ±  0%   14.91Ki ±  1%  -81.11% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=10240-16     89.97Ki ±  0%   25.14Ki ±  1%  -72.06% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=512000-16   1032.8Ki ±  5%   939.1Ki ±  4%   -9.07% (p=0.002 n=6)
geomean                                          258.9Ki         94.02Ki        -63.68%

                                              │  /tmp/old  │             /tmp/new             │
                                              │ allocs/op  │ allocs/op   vs base              │
RPCCompressor/comp=gzip,payloadSize=1024-16     257.0 ± 0%   258.0 ± 0%       ~ (p=0.061 n=6)
RPCCompressor/comp=gzip,payloadSize=10240-16    259.0 ± 0%   258.5 ± 0%       ~ (p=0.182 n=6)
RPCCompressor/comp=gzip,payloadSize=512000-16   305.0 ± 2%   308.0 ± 0%       ~ (p=0.063 n=6)
RPCCompressor/comp=noop,payloadSize=1024-16     234.0 ± 0%   231.0 ± 0%  -1.28% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=10240-16    235.0 ± 0%   231.0 ± 0%  -1.70% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=512000-16   308.0 ± 2%   308.5 ± 3%       ~ (p=0.671 n=6)
geomean                                         264.7        263.9       -0.28%
```