@easwars 2024-10-29
We have another ongoing PR which addresses this issue. Please see: https://github.com/grpc/grpc-go/pull/7653