@easwars 2024-10-29
I think we can close this one in favor of https://github.com/grpc/grpc-go/pull/7653. 

@alanprot Please let us know if you think otherwise.