@PapaCharlie 2024-10-30
@alanprot Can you try replacing the `io.ReadAll` with `mem.ReadAll` (from #7653) in `decompress` and running your benchmark again? I'm interested in seeing the results. Also your benchmark results are interesting:
```
                                              │  /tmp/old   │              /tmp/new              │
                                              │   sec/op    │   sec/op     vs base               │
RPCCompressor/comp=gzip,payloadSize=1024-16     342.0µ ± 3%   295.0µ ± 2%  -13.75% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=10240-16    440.1µ ± 2%   393.9µ ± 2%  -10.49% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=512000-16   5.111m ± 1%   5.179m ± 2%   +1.33% (p=0.004 n=6)
RPCCompressor/comp=noop,payloadSize=1024-16     166.9µ ± 1%   135.3µ ± 1%  -18.93% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=10240-16    183.5µ ± 1%   152.5µ ± 3%  -16.92% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=512000-16   977.7µ ± 2%   941.9µ ± 1%   -3.66% (p=0.002 n=6)
geomean                                         533.4µ        476.4µ       -10.69%
```
I would have expected to see no difference between old and new `noop` tests, but I guess because you're using a buffer from the pool when calling `io.CopyBuffer` instead of letting `io.Copy` create a new one we're seeing some noticeable gains? Nice