@alanprot 2024-11-08
@PapaCharlie 
This are the results with the `mem.ReadAll` from that PR. Seems that indeed that PR fix the issue as well..

```
goos: darwin
goarch: arm64
pkg: google.golang.org/grpc
cpu: Apple M1 Pro
                                              │   /tmp/old   │              /tmp/new              │
                                              │    sec/op    │   sec/op     vs base               │
RPCCompressor/comp=gzip,payloadSize=1024-10     156.5µ ±  2%   146.4µ ± 4%   -6.44% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=10240-10    189.2µ ±  7%   172.2µ ± 3%   -8.99% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=512000-10   1.575m ± 10%   1.540m ± 0%   -2.24% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=1024-10     91.15µ ±  4%   73.58µ ± 0%  -19.27% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=10240-10    97.73µ ±  3%   80.58µ ± 3%  -17.55% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=512000-10   457.6µ ±  2%   463.3µ ± 2%   +1.25% (p=0.041 n=6)
geomean                                         239.8µ         217.8µ        -9.18%

                                              │   /tmp/old    │               /tmp/new               │
                                              │     B/op      │     B/op       vs base               │
RPCCompressor/comp=gzip,payloadSize=1024-10     134.85Ki ± 7%   24.37Ki ±  9%  -81.93% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=10240-10    150.12Ki ± 7%   37.74Ki ± 13%  -74.86% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=512000-10   1144.9Ki ± 5%   987.6Ki ±  6%  -13.74% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=1024-10      78.28Ki ± 0%   14.02Ki ±  1%  -82.09% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=10240-10     88.82Ki ± 0%   23.63Ki ±  1%  -73.39% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=512000-10    975.1Ki ± 3%   857.0Ki ±  3%  -12.12% (p=0.002 n=6)
geomean                                          232.3Ki        79.78Ki        -65.65%

                                              │  /tmp/old  │             /tmp/new             │
                                              │ allocs/op  │ allocs/op   vs base              │
RPCCompressor/comp=gzip,payloadSize=1024-10     252.0 ± 0%   244.0 ± 0%  -3.17% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=10240-10    252.0 ± 0%   244.0 ± 0%  -3.17% (p=0.002 n=6)
RPCCompressor/comp=gzip,payloadSize=512000-10   296.0 ± 1%   288.0 ± 1%  -2.70% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=1024-10     230.0 ± 0%   223.0 ± 0%  -3.04% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=10240-10    230.0 ± 0%   223.0 ± 0%  -3.04% (p=0.002 n=6)
RPCCompressor/comp=noop,payloadSize=512000-10   291.5 ± 2%   279.0 ± 2%  -4.29% (p=0.002 n=6)
geomean                                         257.3        248.9       -3.24%
```