@PapaCharlie 2024-11-08
@alanprot just to be clear, are those benchmark results just with `mem.ReadAll`, *without* your changes? Or do they include your changes as well?