@PapaCharlie 2024-11-08
Nevermind, you can't really use both `mem.ReadAll` and `DecompressedSize` anyway. If #7653 fixes the problem, then I think we can nuke the entire TODO block altogether