@easwars 2024-11-18
Closing this as https://github.com/grpc/grpc-go/pull/7653 has been merged.

Thank you for your contribution though @alanprot