@alanprot 2024-11-28
@easwars There is any timeline to release a new version with https://github.com/grpc/grpc-go/pull/7653???