@purnesh42H 2024-11-29
> @easwars There is any timeline to release a new version with #7653?

It should be released as part of 1.69